✨ Output markdown!
