+++ components/CodeBlock.tsx
@@ -1,15 +1,3 @@
----
-title: "Two Forms of Pre-rendering"
-date: "2020-01-01"
----
-
-# Ahiahi
-
-でコードをシンタックスハイライトさせるためには、カスタムコンポーネントを作成する必要があります。以下のリンク先に書いてあるようにコンポーネントを作成していきます。
-
-これが`ahiahi`です。
-
-```tsx
 import { CodeComponent } from "react-markdown/src/ast-to-react";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import { dracula } from "react-syntax-highlighter/dist/cjs/styles/prism";
@@ -29,6 +17,3 @@ const CodeBlock: CodeComponent = ({ node, inline, className, children }) => {
 };
 
 export default CodeBlock;
-```
-
-Importantly, Next.js lets you **choose** which pre-rendering form to use for each page. You can create a "hybrid" Next.js app by using Static Generation for most pages and using Server-side Rendering for others.
