--- components/markdown.tsx
@@ -1,49 +1,0 @@
-import ReactMarkdown from "react-markdown";
-import {
-  CodeComponent,
-  ReactMarkdownNames,
-} from "react-markdown/src/ast-to-react";
-import { materialLight } from "react-syntax-highlighter/dist/cjs/styles/prism";
-import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
-import { FunctionComponent } from "react";
-
-interface IProps {
-  content: string;
-}
-
-const Markdown: FunctionComponent<IProps> = ({ content }) => {
-  const components = {
-    code: CodeBlock,
-  };
-
-  return (
-    <div className="markdown-body">
-      <ReactMarkdown components={components}>{content}</ReactMarkdown>
-    </div>
-  );
-};
-
-const CodeBlock: CodeComponent | ReactMarkdownNames = ({
-  inline,
-  className,
-  children,
-  ...props
-}) => {
-  const match = /language-(\w+)/.exec(className || "");
-  return !inline && match ? (
-    <SyntaxHighlighter
-      style={materialLight}
-      language={match[1]}
-      PreTag="div"
-      {...props}
-    >
-      {String(children).replace(/\n$/, "")}
-    </SyntaxHighlighter>
-  ) : (
-    <code className={className ? className : ""} {...props}>
-      {children}
-    </code>
-  );
-};
-
-export default Markdown;
